+++ apps/docs/.storybook/manager.tsx
@@ -2,9 +2,10 @@ import React from "react";
 import { addons, types } from "@storybook/manager-api";
 import { IconButton } from "@storybook/components";
 import { HomeIcon } from "@storybook/icons";
+import { STORY_RENDERED } from "@storybook/core-events";
 
-addons.register("home", () => {
-  addons.add("home/tool", {
+addons.register("custom", (api) => {
+  addons.add("custom/home", {
     type: types.TOOL,
     title: "Home",
     render: () => (
@@ -16,5 +17,11 @@ addons.register("home", () => {
         <HomeIcon />
       </IconButton>
     ),
+  });
+
+  api.on(STORY_RENDERED, () => {
+    const story = api.getCurrentStoryData();
+
+    document.title = `${story.title.replace("/", " / ")} ⋅ Yamori Design`;
   });
 });