+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @yamori-design/icons
 
+## 1.0.0
+
+### Major Changes
+
+- Finalize major structure with major release
+
 ## 0.1.0
 
 ### Minor Changes